+++ src/main/java/main/repositories/CardRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CardRepository extends CrudRepository<Card, Long> {
+public interface CardRepository extends CrudRepository<Card, Integer> {
     Card findByName (String name);
     Card findByCubeNumber (int number);
     void deleteByName (String name);
