+++ src/main/java/main/repositories/CharacterRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CharacterRepository extends CrudRepository<Character, Long> {
+public interface CharacterRepository extends CrudRepository<Character, Integer> {
     Character findByName(String name);
 
     void deleteByName (String name);
