+++ src/main/java/main/repositories/CreatureRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CreatureRepository  extends CrudRepository<Creature, Long>{
+public interface CreatureRepository  extends CrudRepository<Creature, Integer>{
     Creature findByName(String name);
 }
