+++ src/main/java/main/repositories/EpisodeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EpisodeRepository extends CrudRepository<Episode, Long> {
+public interface EpisodeRepository extends CrudRepository<Episode, Integer> {
     Episode findBySeries(int series);
     void deleteBySeries(int series);
 }
