+++ src/main/java/main/repositories/FeatureRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FeatureRepository extends CrudRepository<Feature, Long> {
+public interface FeatureRepository extends CrudRepository<Feature, Integer> {
 }
