+++ src/main/java/main/repositories/NewsRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Date;
 
 @Repository
-public interface NewsRepository extends CrudRepository<News, Long> {
+public interface NewsRepository extends CrudRepository<News, Integer> {
     Iterable<News> getAllByTopic(String topic);
     Iterable<News> getAllByDateAfter(Date after);
     void deleteAllByTopic(String topic);
